backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiohttp>=3.11.0
websockets>=14.0
Pillow>=11.0.0
python-multipart>=0.0.18
